only a few specs to fix
