+++ lib/gantree.rb
@@ -3,6 +3,7 @@ require "gantree/base"
 require "gantree/version"
 require "gantree/deploy"
 require "gantree/deploy_applications"
+require "gantree/deploy_version"
 require "gantree/init"
 require "gantree/delete"
 require "gantree/update"
